requirements.txt
streamlit
yfinance
fpdf2
plotly
kaleido
groq
python-dotenv